requirements.txt
absl-py==2.1.0
accelerate==1.12.0
affine==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aiosignal==1.4.0
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
appnope==0.1.4
APScheduler==3.11.2
argcomplete==3.6.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
ascii==3.6
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==25.4.0
babel==2.16.0
beautifulsoup4==4.12.3
binom_probability==1.3
bleach==6.1.0
blinker==1.9.0
boto3==1.37.8
boto3-session==0.0.8
botocore==1.37.8
braceexpand==0.1.7
branca==0.8.2
cachetools==5.5.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.7
click-plugins==1.1.1.2
cliez==2.1.1
cligj==0.7.2
cloudpickle==3.1.2
cmdstanpy==1.2.4
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.8.2
comb==0.9.10
comm==0.2.2
contourpy==1.3.0
cryptography==46.0.3
cycler==0.12.1
datasets==4.4.1
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
diffusers==0.32.2
dill==0.4.0
distro==1.9.0
dm-tree==0.1.9
docstring_parser==0.17.0
einops==0.8.1
et_xmlfile==2.0.0
etils==1.13.0
evaluate==0.4.6
evo==1.31.1
executing==2.0.1
fastapi==0.121.2
fastjsonschema==2.20.0
filelock==3.20.0
flatbuffers==24.3.25
folium==0.20.0
fonttools==4.53.1
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.10.0
funcsigs==1.0.2
future==1.0.0
gast==0.6.0
gdown==5.2.0
gensim==4.3.3
geodatasets==2024.8.0
geopandas==1.1.1
gin-config==0.5.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
googleapis-common-protos==1.72.0
grpcio==1.66.1
gym==0.26.2
gym-notices==0.1.0
h11==0.16.0
h5py==3.11.0
hf-xet==1.2.0
holidays==0.62
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
imageio==2.37.0
imgviz==1.7.5
immutabledict==4.2.2
importlib_metadata==8.6.1
importlib_resources==6.4.5
imshow==0.2.3.post0
instaloader==4.14.2
ipykernel==6.29.5
ipython==8.26.0
ipywidgets==8.1.8
isoduration==20.11.0
jax==0.8.0
jaxlib==0.8.0
jedi==0.19.1
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.4.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.16
kaggle==1.7.4.5
kiwisolver==1.4.7
labelme==5.9.0
lazy_loader==0.4
libclang==18.1.1
loguru==0.7.3
lz4==4.4.4
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml_dtypes==0.5.3
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
namex==0.0.8
narwhals==1.35.0
natsort==8.4.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
network==0.1
networkx==3.5
nltk==3.9.2
notebook==7.4.1
notebook_shim==0.2.4
numexpr==2.10.2
numpy==2.3.5
onnxruntime==1.23.1
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
openpyxl==3.1.5
opt-einsum==3.3.0
optree==0.12.1
osam==0.2.5
osmnx==2.0.6
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
pooch==1.8.2
prometheus_client==0.20.0
promise==2.3
prompt_toolkit==3.0.47
propcache==0.4.1
prophet==1.1.6
protobuf==4.25.4
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==22.0.0
pycparser==2.22
pydantic==2.12.2
pydantic_core==2.41.4
pydeck==0.9.1
pyglet==2.0.17
Pygments==2.18.0
PyJWT==2.10.1
pyogrio==0.10.0
pyparsing==3.1.4
PyPDF2==3.0.1
pyproj==3.7.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.17
PyQt5_sip==12.17.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-http-client==3.3.7
python-json-logger==2.0.7
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.3
pyzmq==26.2.0
rasterio==1.4.3
redis==7.0.1
referencing==0.35.1
regex==2025.11.3
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.8.0
rosbags==0.10.9
rpds-py==0.20.0
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
s3transfer==0.11.4
safetensors==0.7.0
scikit-image==0.25.2
scikit-learn==1.5.1
scipy==1.16.3
seaborn==0.13.2
Send2Trash==1.8.3
sendgrid==6.12.5
sentence-transformers==5.1.2
seqeval==1.2.2
setuptools==80.9.0
shapely==2.1.2
simple-parsing==0.1.7
six==1.17.0
smart_open==7.3.1
smmap==5.0.2
sniffio==1.3.1
Solution==5.4.0
soupsieve==2.6
SQLAlchemy==2.0.45
stack-data==0.6.3
stanio==0.5.1
starlette==0.49.3
streamlit==1.44.1
symbol_please==1.2.0
sympy==1.13.1
tabula-py==2.10.0
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow-datasets==4.9.9
tensorflow-metadata==1.17.2
tensorflow-text==2.19.0
termcolor==2.4.0
terminado==0.18.1
text-unidecode==1.3
threadpoolctl==3.5.0
tifffile==2025.10.16
tinycss2==1.3.0
tokenizers==0.22.1
toml==0.10.2
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
trax==1.4.1
twilio==9.9.1
types-python-dateutil==2.9.0.20240906
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.38.0
wcwidth==0.2.13
webcolors==24.8.0
webdataset==1.0.2
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.4
wheel==0.44.0
widgetsnbextension==4.0.14
wrapt==1.16.0
xxhash==3.6.0
xyzservices==2025.4.0
yarl==1.22.0
zipp==3.21.0
zstandard==0.23.0
